dfsf

+++ src/main/java/team/bahor/BoomAcademyApplication.java
@@ -5,15 +5,11 @@ import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.context.annotation.Bean;
 import org.springframework.security.crypto.password.PasswordEncoder;
-import team.bahor.entity.base.Auditable;
 import team.bahor.entity.user.AuthUser;
 import team.bahor.enums.Role;
-import team.bahor.property.OpenApiProperty;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import team.bahor.properties.ServerProperties;
+import team.bahor.property.OpenApiProperty;
 import team.bahor.repositories.auth.AuthUserRepository;
 
 import java.util.UUID;
@@ -48,7 +44,7 @@ public class BoomAcademyApplication {
                     .username("admin")
                     .password(encode)
                     .role(Role.ADMIN)
-                    .status(0)
+                    .status((short)0)
                     .fullName("adminov admin adminovich")
                     .phone("+998906543210")
                     .balance(0D)
